Ignore event PATCH errors
Events on calendars have been added by many people.  Often, even though the calendar is permitted to be edited by Systems, the individual events are locked down.

The Gateway code had been updated to quietly log any errors, but this only works with a logger enabled, which we are not using.

This update ignores all PATCH errors.

Updates #27